+++ game-systemshock225thanniversaryremaster/index.js
@@ -2,8 +2,8 @@
 Name: System Shock 2 (Classic AND 25th Anniversary Remaster) Vortex Extension
 Structure: Basic game w/ mods folder
 Author: ChemBoy1
-Version: 0.3.0
-Date: 2025-07-01
+Version: 0.4.0
+Date: 2025-07-02
 ////////////////////////////////////////////////*/
 
 //Import libraries
@@ -26,13 +26,18 @@ const XBOXAPP_ID = "";
 const XBOXEXECNAME = "";
 const GAME_ID = "systemshock225thanniversaryremaster";
 const GAME_ID_CLASSIC = 'systemshock2'
-const GAME_NAME = "System Shock 2";
-const GAME_NAME_SHORT = "System Shock 2";
+const GAME_NAME = "System Shock 2: 25th Anniversary Remaster";
+const GAME_NAME_SHORT = "SS2 Remaster";
+const GAME_NAME_CLASSIC = "System Shock 2 (1999)";
+const GAME_NAME_SHORT_CLASSIC = "System Shock 2";
 let GAME_PATH = undefined;
+let GAME_PATH_CLASSIC = undefined;
 let GAME_VERSION = '';
-let GAME_RELEASE = '';
+let GAME_VERSION_CLASSIC = '';
 let STAGING_FOLDER = '';
+let STAGING_FOLDER_CLASSIC = '';
 let DOWNLOAD_FOLDER = '';
+let DOWNLOAD_FOLDER_CLASSIC = '';
 
 //Different exes for each version, use getExecutable function in context.registerGame
 const EXEC_STEAM = 'hathor_Shipping_Playfab_Steam_x64.exe';
@@ -40,8 +45,9 @@ const EXEC_GOG = 'hathor_Shipping_Playfab_Galaxy_x64.exe';
 const EXEC_EPIC = 'hathor_Shipping_Playfab_Epic_x64.exe'; //NOT SURE IF THIS IS CORRECT. Need a user with Epic version to confirm.
 const EXEC_CLASSIC = 'SS2.exe';
 
-//Info for mod types and installers
+//For Remaster (2025) version
 const USER_HOME = util.getVortexPath('home');
+const CONFIG_PATH = path.join(USER_HOME, 'Saved Games', 'Nightdive Studios', 'System Shock 2 Remastered');
 
 const ROOT_ID = `${GAME_ID}-root`;
 const ROOT_NAME = "Binaries / Root Folder";
@@ -57,28 +63,28 @@ const LEGACY_NAME = "Converted Legacy Mod";
 const LEGACY_PATH = MOD_PATH;
 const LEGACY_FOLDERS = ['obj', 'mesh', 'bitmap', 'motions', 'sq_scripts', 'sdn2', 'strings', 'iface', 'intrface', 'misdml', 'snd']; //cannot put "custscenes" here since it would conflict with Root installer
 
-const SS2TOOL_ID = `${GAME_ID}-ss2tool`;
+const REQ_FILE = 'base.kpf';
+
+// For Classic (1999) version
+const SS2TOOL_ID = `${GAME_ID_CLASSIC}-ss2tool`;
 const SS2TOOL_NAME = `SS2Tool`;
 const SS2TOOL_EXEC = `SS2Tool-v.6.1.1.8.exe`;
 const SS2TOOL_URL = `https://www.systemshock.org/index.php?topic=4141.0`;
 const SS2TOOL_URL_DIRECT = `https://www.systemshock.org/index.php?action=dlattach;topic=4141.0;attach=15728&mid=45944&sesc=1dd1fec0fa9126c5ef752f3685775d4ce5f81760`;
 
-const BMM_ID = `${GAME_ID}-bmm`;
+const BMM_ID = `${GAME_ID_CLASSIC}-bmm`;
 const BMM_NAME = `Blue Mod Manager`;
 const BMM_EXEC = 'ss2bmm.exe';
 
-const EDITOR_ID = `${GAME_ID}-editor`;
+const EDITOR_ID = `${GAME_ID_CLASSIC}-editor`;
 const EDITOR_NAME = `ShockEd`;
 const EDITOR_EXEC = 'ShockEd.exe';
 
-const CLASSIC_ID = `${GAME_ID}-classicmod`;
+const CLASSIC_ID = `${GAME_ID_CLASSIC}-classicmod`;
 const CLASSIC_NAME = `Classic Mod`;
 const CLASSIC_PATH = path.join('DMM');
 
-const CONFIG_PATH = path.join(USER_HOME, 'Saved Games', 'Nightdive Studios', 'System Shock 2 Remastered');
-
-//const REQ_FILE = 'base.kpf';
-const REQ_FILE = '';
+const REQ_FILE_CLASSIC = EXEC_CLASSIC;
 
 //Filled in from info above
 const spec = {
@@ -87,11 +93,12 @@ const spec = {
     "name": GAME_NAME,
     "shortName": GAME_NAME_SHORT,
     "logo": `${GAME_ID_CLASSIC}.jpg`,
+    "modPath": MOD_PATH,
     "modPathIsRelative": true,
     "mergeMods": true,
     "requiresCleanup": true,
     "requiredFiles": [
-      //REQ_FILE,
+      REQ_FILE,
     ],
     "details": {
       "steamAppId": STEAMAPP_ID,
@@ -127,25 +134,55 @@ const spec = {
       "priority": "high",
       "targetPath": `{gamePath}`
     },
+  ],
+  "discovery": {
+    "ids": [
+      STEAMAPP_ID,
+      //EPICAPP_ID,
+      GOGAPP_ID,
+      //XBOXAPP_ID
+    ],
+    "names": []
+  }
+};
+
+//Filled in from info above
+const specClassic = {
+  "game": {
+    "id": GAME_ID_CLASSIC,
+    "name": GAME_NAME_CLASSIC,
+    "shortName": GAME_NAME_SHORT_CLASSIC,
+    "executable": EXEC_CLASSIC,
+    "logo": `${GAME_ID_CLASSIC}.jpg`,
+    "modPath": CLASSIC_PATH,
+    "modPathIsRelative": true,
+    "mergeMods": true,
+    "requiresCleanup": true,
+    "requiredFiles": [
+      REQ_FILE_CLASSIC,
+    ],
+    "details": {
+      "steamAppId": STEAMAPP_ID_CLASSIC,
+      "gogAppId": GOGAPP_ID_CLASSIC,
+      "supportsSymlinks": true,
+      "compatibleDownloads": [GAME_ID],
+    },
+    "environment": {
+      "SteamAPPId": STEAMAPP_ID_CLASSIC,
+      "GogAPPId": GOGAPP_ID_CLASSIC,
+    }
+  },
+  "modTypes": [
     {
       "id": CLASSIC_ID,
       "name": CLASSIC_NAME,
       "priority": "low",
       "targetPath": `{gamePath}\\${CLASSIC_PATH}`
     },
-    {
-      "id": SS2TOOL_ID,
-      "name": SS2TOOL_NAME,
-      "priority": "low",
-      "targetPath": `{gamePath}`
-    },
   ],
   "discovery": {
     "ids": [
-      STEAMAPP_ID,
       STEAMAPP_ID_CLASSIC,
-      //EPICAPP_ID,
-      GOGAPP_ID,
       GOGAPP_ID_CLASSIC,
       //XBOXAPP_ID
     ],
@@ -169,21 +206,16 @@ function pathPattern(api, game, pattern) {
   return template(pattern, {
     gamePath: (_a = api.getState().settings.gameMode.discovered[game.id]) === null || _a === void 0 ? void 0 : _a.path,
     documents: util.getVortexPath('documents'),
-    localAppData: process.env['LOCALAPPDATA'],
+    localAppData: util.getVortexPath('localAppData'),
     appData: util.getVortexPath('appData'),
   });
 }
 
-//Set the mod path for the game
-function makeGetModPath(gamePath) {
-  /*let path = 'mods';
-  try {
-    fsNative.stat(path.join(gamePath, EXEC_CLASSIC));
-    path = CLASSIC_PATH;
-  }
-  catch (err) {}
-  return () => path; //*/
-  return () => '.';
+//Set mod path
+function makeGetModPath(api, gameSpec) {
+  return () => gameSpec.game.modPathIsRelative !== false
+    ? gameSpec.game.modPath || '.'
+    : pathPattern(api, gameSpec.game, gameSpec.game.modPath);
 }
 
 //Find game installation directory
@@ -218,9 +250,9 @@ async function requiresLauncher(gamePath, store) {
   return Promise.resolve(undefined);
 }
 
-const getDiscoveryPath = (api) => {
+const getDiscoveryPath = (api, gameSpec) => {
   const state = api.getState();
-  const discovery = util.getSafe(state, [`settings`, `gameMode`, `discovered`, GAME_ID], {});
+  const discovery = util.getSafe(state, [`settings`, `gameMode`, `discovered`, gameSpec.game.id], {});
   return discovery === null || discovery === void 0 ? void 0 : discovery.path;
 };
 
@@ -269,13 +301,13 @@ async function browseForDownloadFunction(api, gameSpec, URL, instructions, ARCHI
               if (error !== null) {
                 return reject(error);
               }
-              const profileId = selectors.lastActiveProfileForGame(api.getState(), GAME_ID);
+              const profileId = selectors.lastActiveProfileForGame(api.getState(), gameSpec.game.id);
               const batched = [
                 actions.setModsEnabled(api, profileId, result, true, {
                   allowAutoDeploy: true,
                   installed: true,
                 }),
-                actions.setModType(GAME_ID, result[0], MOD_TYPE), // Set the mod type
+                actions.setModType(gameSpec.game.id, result[0], MOD_TYPE), // Set the mod type
               ];
               util.batchDispatch(api.store, batched); // Will dispatch both actions.
               try {
@@ -335,13 +367,13 @@ async function browseForDownloadFunction(api, gameSpec, URL, instructions, ARCHI
               if (error !== null) {
                 return reject(error);
               }
-              const profileId = selectors.lastActiveProfileForGame(api.getState(), GAME_ID);
+              const profileId = selectors.lastActiveProfileForGame(api.getState(), gameSpec.game.id);
               const batched = [
                 actions.setModsEnabled(api, profileId, result, true, {
                   allowAutoDeploy: true,
                   installed: true,
                 }),
-                actions.setModType(GAME_ID, result[0], MOD_TYPE), // Set the mod type
+                actions.setModType(gameSpec.game.id, result[0], MOD_TYPE), // Set the mod type
               ];
               util.batchDispatch(api.store, batched); // Will dispatch both actions.
               try { //run installer from staging folder with elevation
@@ -499,13 +531,13 @@ async function browseForDownloadFunction(api, gameSpec, URL, instructions, ARCHI
               if (error !== null) {
                 return reject(error);
               }
-              const profileId = selectors.lastActiveProfileForGame(api.getState(), GAME_ID);
+              const profileId = selectors.lastActiveProfileForGame(api.getState(), gameSpec.game.id);
               const batched = [
                 actions.setModsEnabled(api, profileId, result, true, {
                   allowAutoDeploy: true,
                   installed: true,
                 }),
-                actions.setModType(GAME_ID, result[0], MOD_TYPE), // Set the mod type
+                actions.setModType(gameSpec.game.id, result[0], MOD_TYPE), // Set the mod type
               ];
               util.batchDispatch(api.store, batched); // Will dispatch both actions.
               return resolve();
@@ -532,8 +564,8 @@ async function browseForDownloadFunction(api, gameSpec, URL, instructions, ARCHI
 } //*/
 
 //Check if SS2Tool is installed
-function isSS2ToolInstalled(api) {
-  GAME_PATH = getDiscoveryPath(api);
+function isSS2ToolInstalled(api, gameSpec) {
+  GAME_PATH = getDiscoveryPath(api, gameSpec);
   try {
     fs.statSync(path.join(GAME_PATH, BMM_EXEC));
     return true;
@@ -544,7 +576,7 @@ function isSS2ToolInstalled(api) {
 
 //* Function to have user browse to download RS_ASIO
 async function downloadSS2Tool(api, gameSpec) {
-  let isInstalled = isSS2ToolInstalled(api);
+  let isInstalled = isSS2ToolInstalled(api, gameSpec);
   const URL = SS2TOOL_URL;
   const MOD_NAME = SS2TOOL_NAME;
   const MOD_TYPE = SS2TOOL_ID;
@@ -561,13 +593,12 @@ async function downloadSS2Tool(api, gameSpec) {
   await browseForDownloadFunction(api, gameSpec, URL, instructions, ARCHIVE_NAME, MOD_NAME, isArchive, isInstaller, INSTALLER, STAGING_PATH, MOD_TYPE, isInstalled, isElevated);
 }
 
-// MOD INSTALLER FUNCTIONS ///////////////////////////////////////////////////
+// REMASTER MOD INSTALLER FUNCTIONS ///////////////////////////////////////////////////
 
 //Test for save files
 function testMod(files, gameId) {
   const isMod = files.some(file => MOD_EXT.includes(path.extname(file).toLowerCase()))
-  const isClassic = GAME_RELEASE === 'classic';
-  let supported = (gameId === spec.game.id) && isMod && !isClassic;
+  let supported = (gameId === spec.game.id) && isMod;
 
   // Test for a mod installer
   if (supported && files.find(file =>
@@ -612,7 +643,7 @@ function toBlue(func) {
 
 //Success notifications
 function convertSuccessNotify(api, name) {
-  const NOTIF_ID = `${GAME_ID}-folonlinksuccess`;
+  const NOTIF_ID = `${GAME_ID}-legacyconvertsuccess`;
   const MESSAGE = `Successfully converted legacy SS2 Mod "${name}" to .kpf format.`;
   api.sendNotification({
     id: NOTIF_ID,
@@ -626,8 +657,7 @@ function convertSuccessNotify(api, name) {
 //Test for legacy SS2 mod files
 function testLegacy(files, gameId) {
   const isMod = files.some(file => LEGACY_FOLDERS.includes(path.basename(file).toLowerCase()))
-  const isClassic = GAME_RELEASE === 'classic';
-  let supported = (gameId === spec.game.id) && isMod && !isClassic;
+  let supported = (gameId === spec.game.id) && isMod;
 
   // Test for a mod installer
   if (supported && files.find(file =>
@@ -693,8 +723,7 @@ async function installLegacy(files, destinationPath) {
 //Test for save files
 function testRootFolder(files, gameId) {
   const isMod = files.some(file => ROOT_FOLDERS.includes(path.basename(file).toLowerCase()))
-  const isClassic = GAME_RELEASE === 'classic';
-  let supported = (gameId === spec.game.id) && !isClassic;
+  let supported = (gameId === spec.game.id) && isMod;
 
   // Test for a mod installer
   if (supported && files.find(file =>
@@ -734,8 +763,7 @@ function installRootFolder(files) {
 
 //Test for fallback binaries installer
 function testRoot(files, gameId) {
-  const isClassic = GAME_RELEASE === 'classic';
-  let supported = (gameId === spec.game.id) && !isClassic;
+  let supported = (gameId === spec.game.id);
 
   // Test for a mod installer.
   if (supported && files.find(file =>
@@ -768,11 +796,12 @@ function installRoot(files) {
   return Promise.resolve({ instructions });
 }
 
+// CLASSIC MOD INSTALLER FUNCTIONS ///////////////////////////////////////////////////
+
 //Test for fallback binaries installer
 function testClassic(files, gameId) {
   const isMod = files.some(file => LEGACY_FOLDERS.includes(path.basename(file).toLowerCase()))
-  const isClassic = GAME_RELEASE === 'classic';
-  let supported = (gameId === spec.game.id) && isMod && isClassic;
+  let supported = (gameId === specClassic.game.id) && isMod;
 
   // Test for a mod installer.
   if (supported && files.find(file =>
@@ -850,16 +879,12 @@ function getExecutable(discoveryPath) {
     GAME_VERSION = 'epic';
     return EXEC_EPIC;
   };
-  if (isCorrectExec(EXEC_CLASSIC)) {
-    GAME_VERSION = 'classic';
-    return EXEC_CLASSIC;
-  };
   return EXEC_STEAM;
 }
 
 //Get game store version based on which executable is in the game folder
-function getVersion(api) {
-  GAME_PATH = getDiscoveryPath(api);
+function getVersion(api, gameSpec) {
+  GAME_PATH = getDiscoveryPath(api, gameSpec);
   const isCorrectExec = (exec) => {
     try {
       fs.statSync(path.join(GAME_PATH, exec));
@@ -881,41 +906,6 @@ function getVersion(api) {
     GAME_VERSION = 'epic';
     return GAME_VERSION;
   };
-  if (isCorrectExec(EXEC_CLASSIC)) {
-    GAME_VERSION = 'classic';
-    return GAME_VERSION;
-  };
-  return;
-}
-
-//Get game release version (classic or remaster) based on which executable is in the game folder
-function getRelease(api) {
-  GAME_PATH = getDiscoveryPath(api);
-  const isCorrectExec = (exec) => {
-    try {
-      fs.statSync(path.join(GAME_PATH, exec));
-      return true;
-    }
-    catch (err) {
-      return false;
-    }
-  };
-  if (isCorrectExec(EXEC_STEAM)) {
-    GAME_RELEASE = 'remaster';
-    return GAME_RELEASE;
-  };
-  if (isCorrectExec(EXEC_GOG)) {
-    GAME_RELEASE = 'remaster';
-    return GAME_RELEASE;
-  };
-  if (isCorrectExec(EXEC_EPIC)) {
-    GAME_RELEASE = 'remaster';
-    return GAME_RELEASE;
-  };
-  if (isCorrectExec(EXEC_CLASSIC)) {
-    GAME_RELEASE = 'classic';
-    return GAME_RELEASE;
-  };
   return;
 }
 
@@ -923,19 +913,20 @@ function getRelease(api) {
 async function setup(discovery, api, gameSpec) {
   const state = api.getState();
   GAME_PATH = discovery.path;
-  GAME_VERSION = getVersion(api);
-  GAME_RELEASE = getRelease(api);
-  STAGING_FOLDER = selectors.installPathForGame(state, GAME_ID);
-  DOWNLOAD_FOLDER = selectors.downloadPathForGame(state, GAME_ID);
-  if (GAME_RELEASE === 'classic') {
-    try {
-      await downloadSS2Tool(api, gameSpec);
-    } catch (err) {
-      log('error', `Failed to download SS2Tool: ${err}`);
-    }
-  }
-  await fs.ensureDirWritableAsync(path.join(GAME_PATH, CLASSIC_PATH));
+  GAME_VERSION = getVersion(api, gameSpec);
+  STAGING_FOLDER = selectors.installPathForGame(state, gameSpec.game.id);
+  DOWNLOAD_FOLDER = selectors.downloadPathForGame(state, gameSpec.game.id);
   return fs.ensureDirWritableAsync(path.join(GAME_PATH, MOD_PATH));
+}
+
+//Setup function
+async function setupClassic(discovery, api, gameSpec) {
+  const state = api.getState();
+  GAME_PATH_CLASSIC = discovery.path;
+  STAGING_FOLDER_CLASSIC = selectors.installPathForGame(state, gameSpec.game.id);
+  DOWNLOAD_FOLDER_CLASSIC = selectors.downloadPathForGame(state, gameSpec.game.id);
+  await downloadSS2Tool(api, gameSpec);
+  return fs.ensureDirWritableAsync(path.join(GAME_PATH_CLASSIC, CLASSIC_PATH));
 }
 
 //Let Vortex know about the game
@@ -961,6 +952,76 @@ function applyGame(context, gameSpec) {
         shell: true,
         parameters: []
       }, //*/
+    ],
+  };
+  context.registerGame(game);
+
+  //register mod types
+  (gameSpec.modTypes || []).forEach((type, idx) => {
+    context.registerModType(type.id, modTypePriority(type.priority) + idx, (gameId) => {
+      var _a;
+      return (gameId === gameSpec.game.id)
+        && !!((_a = context.api.getState().settings.gameMode.discovered[gameId]) === null || _a === void 0 ? void 0 : _a.path);
+    }, (game) => pathPattern(context.api, game, type.targetPath), () => Promise.resolve(false), { name: type.name });
+  });
+
+  //register mod installers
+  context.registerInstaller(MOD_ID, 25, testMod, installMod);
+  context.registerInstaller(LEGACY_ID, 27, toBlue(testLegacy), toBlue(installLegacy));
+  //context.registerInstaller(LEGACY_ID, 27, toBlue(testLegacy), (files, destinationPath) => toBlue(installLegacy(context.api, files, destinationPath))); //with api passed in
+  context.registerInstaller(`${ROOT_ID}folder`, 29, testRootFolder, installRootFolder);
+  context.registerInstaller(ROOT_ID, 31, testRoot, installRoot); //fallback to root folder
+
+  //register actions
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open Config/Save Folder', () => {
+    const openPath = CONFIG_PATH;
+    util.opn(openPath).catch(() => null);
+    }, () => {
+      const state = context.api.getState();
+      const gameId = selectors.activeGameId(state);
+      return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'View Changelog', () => {
+    const openPath = path.join(__dirname, 'CHANGELOG.md');
+    util.opn(openPath).catch(() => null);
+    }, () => {
+      const state = context.api.getState();
+      const gameId = selectors.activeGameId(state);
+      return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open Downloads Folder', () => {
+    const openPath = DOWNLOAD_FOLDER;
+    util.opn(openPath).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+}
+
+//Let Vortex know about the game
+function applyGameClassic(context, gameSpec) {
+  //register game
+  const game = {
+    ...gameSpec.game,
+    queryPath: makeFindGame(context.api, gameSpec),
+    queryModPath: makeGetModPath(context.api, gameSpec),
+    requiresLauncher: requiresLauncher,
+    setup: async (discovery) => await setupClassic(discovery, context.api, gameSpec), //setup function
+    executable: () => gameSpec.game.executable,
+    supportedTools: [
+      {
+        id: `${GAME_ID_CLASSIC}-customlaunch`,
+        name: `Custom Launch`,
+        logo: `exec.png`,
+        executable: () => gameSpec.game.executable,
+        requiredFiles: [gameSpec.game.executable],
+        detach: true,
+        relative: true,
+        exclusive: true,
+        shell: true,
+        parameters: []
+      }, //*/
       {
         id: BMM_ID,
         name: BMM_NAME,
@@ -978,7 +1039,7 @@ function applyGame(context, gameSpec) {
         id: SS2TOOL_ID,
         name: SS2TOOL_NAME,
         logo: `ss2tool.png`,
-        queryPath: () => path.join(DOWNLOAD_FOLDER, SS2TOOL_EXEC),
+        queryPath: () => path.join(DOWNLOAD_FOLDER_CLASSIC, SS2TOOL_EXEC),
         executable: () => SS2TOOL_EXEC,
         requiredFiles: [SS2TOOL_EXEC],
         detach: true,
@@ -1013,21 +1074,15 @@ function applyGame(context, gameSpec) {
   });
 
   //register mod installers
-  context.registerInstaller(MOD_ID, 25, testMod, installMod);
-  context.registerInstaller(LEGACY_ID, 27, toBlue(testLegacy), toBlue(installLegacy));
-  //context.registerInstaller(LEGACY_ID, 27, toBlue(testLegacy), (files, destinationPath) => toBlue(installLegacy(context.api, files, destinationPath)));
-  context.registerInstaller(`${ROOT_ID}folder`, 29, testRootFolder, installRootFolder);
-  context.registerInstaller(ROOT_ID, 31, testRoot, installRoot); //fallback to root folder
   context.registerInstaller(CLASSIC_ID, 33, testClassic, installClassic); //fallback to root folder
 
   //register actions
-  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open Config/Save Folder (Remaster)', () => {
-    const openPath = CONFIG_PATH;
-    util.opn(openPath).catch(() => null);
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Download and Run SS2Tool', () => {
+    downloadSS2Tool(context.api, gameSpec);
     }, () => {
       const state = context.api.getState();
       const gameId = selectors.activeGameId(state);
-      return gameId === GAME_ID;
+      return gameId === GAME_ID_CLASSIC;
   });
   context.registerAction('mod-icons', 300, 'open-ext', {}, 'View Changelog', () => {
     const openPath = path.join(__dirname, 'CHANGELOG.md');
@@ -1035,21 +1090,22 @@ function applyGame(context, gameSpec) {
     }, () => {
       const state = context.api.getState();
       const gameId = selectors.activeGameId(state);
-      return gameId === GAME_ID;
+      return gameId === GAME_ID_CLASSIC;
   });
   context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open Downloads Folder', () => {
-    const openPath = DOWNLOAD_FOLDER;
+    const openPath = DOWNLOAD_FOLDER_CLASSIC;
     util.opn(openPath).catch(() => null);
   }, () => {
     const state = context.api.getState();
     const gameId = selectors.activeGameId(state);
-    return gameId === GAME_ID;
+    return gameId === GAME_ID_CLASSIC;
   });
 }
 
 //main function
 function main(context) {
   applyGame(context, spec);
+  applyGameClassic(context, specClassic);
   context.once(() => { // put code here that should be run (once) when Vortex starts up
 
   });